create new component for action bar
